+++ src/main/java/com/rko/huemanager/repository/ScheduleRepository.java
@@ -1,6 +1,7 @@
 package com.rko.huemanager.repository;
 
 import com.rko.huemanager.domain.Schedule;
+import com.rko.huemanager.domain.constant.ScheduleStatus;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -8,5 +9,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.time.LocalDate;
 
 public interface ScheduleRepository extends JpaRepository<Schedule, Long>, ScheduleCustomRepository {
-    Page<Schedule> findByStartDateBetween(LocalDate startDate, LocalDate endDate, Pageable pageable);
+    Page<Schedule> findByStartDateBetweenAndStatus(LocalDate startDate, LocalDate endDate, ScheduleStatus status, Pageable pageable);
+    Page<Schedule> findByStatus(ScheduleStatus status, Pageable pageable);
 }
