+++ src/main/java/com/rko/huemanager/service/ScheduleService.java
@@ -51,8 +51,8 @@ public class ScheduleService {
     }
 
     @Transactional(readOnly = true)
-    public Page<ScheduleDto> getAllSchedules(Pageable pageable) {
-        return scheduleRepository.findAll(pageable).map(ScheduleDto::from);
+    public Page<ScheduleDto> getAllApprovedSchedules(Pageable pageable) {
+        return scheduleRepository.findByStatus(ScheduleStatus.APPROVED, pageable).map(ScheduleDto::from);
     }
 
     @Transactional(readOnly = true)
@@ -63,7 +63,7 @@ public class ScheduleService {
 
     @Transactional(readOnly = true)
     public Page<ScheduleDto> getDaySchedules(LocalDate date, Pageable pageable){
-        Page<Schedule> schedules = scheduleRepository.findByStartDateBetween(date, date, pageable);
+        Page<Schedule> schedules = scheduleRepository.findByStartDateBetweenAndStatus(date, date, ScheduleStatus.APPROVED, pageable);
         return schedules.map(ScheduleDto::from);
     }
 
@@ -71,7 +71,7 @@ public class ScheduleService {
     public Page<ScheduleDto> getWeekSchedules(LocalDate date, Pageable pageable){
         LocalDate startDate = date.with(previousOrSame(DayOfWeek.MONDAY));
         LocalDate endDate = date.with(nextOrSame(DayOfWeek.SUNDAY));
-        Page<Schedule> schedules = scheduleRepository.findByStartDateBetween(startDate, endDate, pageable);
+        Page<Schedule> schedules = scheduleRepository.findByStartDateBetweenAndStatus(startDate, endDate, ScheduleStatus.APPROVED, pageable);
         return schedules.map(ScheduleDto::from);
     }
 
@@ -79,19 +79,7 @@ public class ScheduleService {
     public Page<ScheduleDto> getMonthSchedules(LocalDate date, Pageable pageable){
         LocalDate startDate = date.with(firstDayOfMonth());
         LocalDate endDate = date.with(lastDayOfMonth());
-        Page<Schedule> schedules = scheduleRepository.findByStartDateBetween(startDate, endDate, pageable);
-        return schedules.map(ScheduleDto::from);
-    }
-
-    @Transactional(readOnly = true)
-    public Page<ScheduleDto> searchSchedules(ScheduleSearchRequest request, Pageable pageable){
-        Page<Schedule> schedules = scheduleRepository.findSearchSchedules(
-                request.startDate(),
-                request.endDate(),
-                request.type(),
-                request.status(),
-                pageable);
-
+        Page<Schedule> schedules = scheduleRepository.findByStartDateBetweenAndStatus(startDate, endDate, ScheduleStatus.APPROVED, pageable);
         return schedules.map(ScheduleDto::from);
     }
